+++ include/Physica/Core/Scalar/DiffImpl/SIMDImpl/SIMDImpl.h
@@ -21,13 +21,13 @@
 #include "../SIMD.h"
 
 namespace Physica {
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     SIMD<Diff<T, Mode, Order>, Size>::SIMD(int x) : values(x), grads(0) {}
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     SIMD<Diff<T, Mode, Order>, Size>::SIMD(double x) : values(x), grads(0) {}
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     SIMD<Diff<T, Mode, Order>, Size>::SIMD(Scalar auto x) {
         if constexpr (Diffable<decltype(x)>) {
             values = ValueType(x.value());
@@ -39,49 +39,49 @@ namespace Physica {
         }
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     SIMD<Diff<T, Mode, Order>, Size>::SIMD(ScalarType x, int count) : values(x.value(), count), grads(x.grad()) {}
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     SIMD<Diff<T, Mode, Order>, Size>::SIMD(ValueType values_) : values(std::move(values_)), grads(0) {}
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     SIMD<Diff<T, Mode, Order>, Size>::SIMD(ValueType values_, GradType grads_) : values(std::move(values_)), grads(std::move(grads_)) {}
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     SIMD<Diff<T, Mode, Order>, Size>::SIMD(HalfType a, HalfType b) : values(a.value(), b.value()), grads(a.grad(), b.grad()) {}
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     template<int OtherOrder>
     SIMD<Diff<T, Mode, Order>, Size>::SIMD(const SIMD<Diff<T, Mode, OtherOrder>, Size>& other)
             : values(other.value()), grads(other.grad()) {}
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::operator[](int index) const -> ScalarType {
         return ScalarType(values[index], grads[index]);
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::operator+(const SIMD& other) const -> This {
         return This(values + other.values, grads + other.grads);
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::operator-(const SIMD& other) const -> This {
         return This(values - other.values, grads - other.grads);
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::operator*(const SIMD& x) const -> This {
         return This(values * x.value(), GradType(GradType(x) * grad() + GradType(*this) * x.grad()));
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::operator*(const ScalarType& x) const -> This {
         return This(values * x.value(), GradType(x.template mask<Order - 1>() * grad() + GradType(*this) * x.grad()));
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::operator*(const Scalar auto& x) const -> This {
         if constexpr (Diffable<decltype(x)>)
             return *this * ScalarType(x);
@@ -89,90 +89,90 @@ namespace Physica {
             return This(values * x, grads * x);
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::operator/(const SIMD& x) const -> This {
         const auto x1 = GradType(x);
         const auto v = reciprocal(x1);
         return This(value() * v.value(), GradType((grad() * GradType(x1) - GradType(*this) * x.grad()) * square(v)));
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::operator-() const -> This {
         return This(-values, -grads);
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::reverse(GradType grad) const noexcept -> ValueType {
         grads += grad;
         return values;
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     void SIMD<Diff<T, Mode, Order>, Size>::load(ConstPtrTy p) {
         values.load(p.value_ptr());
         grads.load(p.grad_ptr());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     void SIMD<Diff<T, Mode, Order>, Size>::load_partial(ConstPtrTy p, int n) {
         values.load_partial(p.value_ptr(), n);
         grads.load_partial(p.grad_ptr(), n);
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     void SIMD<Diff<T, Mode, Order>, Size>::store(PtrTy p) const {
         values.store(p.value_ptr());
         grads.store(p.grad_ptr());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     void SIMD<Diff<T, Mode, Order>, Size>::store_partial(PtrTy p, int n) const {
         values.store_partial(p.value_ptr(), n);
         grads.store_partial(p.grad_ptr(), n);
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::cutoff(int count) -> This& {
         values.cutoff(count);
         grads.cutoff(count);
         return *this;
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::swapRealImag() const noexcept -> FullRealType {
         return FullRealType(values.swapRealImag(), grads.swapRealImag());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::permRealImag() const noexcept -> FullRealType {
         return FullRealType(values.permRealImag(), grads.permRealImag());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::sum() const {
         return ScalarType(values.sum(), grads.sum());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::max() const {
         const T x = values.max();
         return ScalarType(x, GradType::select(ValueType(x) == values, grads, GradType(0)).sum());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::min() const {
         const T x = values.min();
         return ScalarType(x, GradType::select(ValueType(x) == values, grads, GradType(0)).sum());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     void SIMD<Diff<T, Mode, Order>, Size>::swap(SIMD& __restrict obj) noexcept {
         assert(this != &obj && "[Error]: Self swap is likely a bug");
         values.swap(obj.values);
         grads.swap(obj.grads);
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::asReal() const noexcept -> FullRealType {
         if constexpr (T::isComplex)
             return FullRealType(values.asReal(), grads.asReal());
@@ -180,22 +180,22 @@ namespace Physica {
             return *this;
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::real() const noexcept -> RealType {
         return RealType(values.real(), grads.real());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::imag() const noexcept -> RealType {
         return RealType(values.imag(), grads.imag());
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::select(BoolSIMDType flags, const SIMD& x, const SIMD& y) -> This {
         return This(ValueType::select(flags, x.value(), y.value()), GradType::select(flags, x.grad(), y.grad()));
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     auto SIMD<Diff<T, Mode, Order>, Size>::asComplex(const FullRealType& reals) -> This {
         This result{};
         result.values = ValueType::asComplex(reals.value());
@@ -203,7 +203,7 @@ namespace Physica {
         return result;
     }
 
-    template<Scalar T, DiffMode Mode, int Order, size_t Size>
+    template<Scalar T, DiffMode Mode, int Order, int Size>
     [[nodiscard]] CoDiff<SIMD<Diff<T, Mode, Order>, Size>> mul_add(
             const SIMD<Diff<T, Mode, Order>, Size>& a,
             const SIMD<Diff<T, Mode, Order>, Size>& b,
